// typechain-types/DIDRegistry.ts
export declare namespace DIDRegistry {
  export type DIDStruct = {
    controller: AddressLike;
    didDocumentHash: string;
    serviceEndpoint: string;
    publicKey: BytesLike;
    exists: boolean;
  };

  export type DIDStructOutput = [
    controller: string,
    didDocumentHash: string,
    serviceEndpoint: string,
    publicKey: string,
    exists: boolean
  ] & {
    controller: string;
    didDocumentHash: string;
    serviceEndpoint: string;
    publicKey: string;
    exists: boolean;
  };
}

export interface DIDRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "exists"
      | "getDID"
      | "registerDID"
      | "rotateController"
      | "rotateKey"
      | "updateServiceEndpoint"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic:
      | "ControllerRotated"
      | "DIDRegistered"
      | "DIDUpdated"
      | "KeyRotated"
  ): EventFragment;

  encodeFunctionData(functionFragment: "exists", values: [string]): string;
  encodeFunctionData(functionFragment: "getDID", values: [string]): string;
  encodeFunctionData(
    functionFragment: "registerDID",
    values: [AddressLike, string]
  ): string;
  encodeFunctionData(
    functionFragment: "rotateController",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "rotateKey",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "updateServiceEndpoint",
    values: [string]
  ): string;

  decodeFunctionResult(functionFragment: "exists", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getDID", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "registerDID",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "rotateController",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "rotateKey", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "updateServiceEndpoint",
    data: BytesLike
  ): Result;
}

export namespace ControllerRotatedEvent {
  export type InputTuple = [
    did: string,
    oldController: AddressLike,
    newController: AddressLike
  ];
  export type OutputTuple = [
    did: string,
    oldController: string,
    newController: string
  ];
  export interface OutputObject {
    did: string;
    oldController: string;
    newController: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace DIDRegisteredEvent {
  export type InputTuple = [
    did: string,
    controller: AddressLike,
    didDocumentHash: string,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    did: string,
    controller: string,
    didDocumentHash: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    did: string;
    controller: string;
    didDocumentHash: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace DIDUpdatedEvent {
  export type InputTuple = [did: string, newEndpoint: string];
  export type OutputTuple = [did: string, newEndpoint: string];
  export interface OutputObject {
    did: string;
    newEndpoint: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace KeyRotatedEvent {
  export type InputTuple = [did: string, oldKey: BytesLike, newKey: BytesLike];
  export type OutputTuple = [did: string, oldKey: string, newKey: string];
  export interface OutputObject {
    did: string;
    oldKey: string;
    newKey: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface DIDRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): DIDRegistry;
  waitForDeployment(): Promise<this>;

  interface: DIDRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  exists: TypedContractMethod<[did: string], [boolean], "view">;

  getDID: TypedContractMethod<
    [did: string],
    [DIDRegistry.DIDStructOutput],
    "view"
  >;

  registerDID: TypedContractMethod<
    [controller: AddressLike, didDocumentHash: string],
    [string],
    "nonpayable"
  >;

  rotateController: TypedContractMethod<
    [newController: AddressLike],
    [string],
    "nonpayable"
  >;

  rotateKey: TypedContractMethod<[newKey: BytesLike], [string], "nonpayable">;

  updateServiceEndpoint: TypedContractMethod<
    [newEndpoint: string],
    [string],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "exists"
  ): TypedContractMethod<[did: string], [boolean], "view">;
  getFunction(
    nameOrSignature: "getDID"
  ): TypedContractMethod<[did: string], [DIDRegistry.DIDStructOutput], "view">;
  getFunction(
    nameOrSignature: "registerDID"
  ): TypedContractMethod<
    [controller: AddressLike, didDocumentHash: string],
    [string],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "rotateController"
  ): TypedContractMethod<[newController: AddressLike], [string], "nonpayable">;
  getFunction(
    nameOrSignature: "rotateKey"
  ): TypedContractMethod<[newKey: BytesLike], [string], "nonpayable">;
  getFunction(
    nameOrSignature: "updateServiceEndpoint"
  ): TypedContractMethod<[newEndpoint: string], [string], "nonpayable">;

  getEvent(
    key: "ControllerRotated"
  ): TypedContractEvent<
    ControllerRotatedEvent.InputTuple,
    ControllerRotatedEvent.OutputTuple,
    ControllerRotatedEvent.OutputObject
  >;
  getEvent(
    key: "DIDRegistered"
  ): TypedContractEvent<
    DIDRegisteredEvent.InputTuple,
    DIDRegisteredEvent.OutputTuple,
    DIDRegisteredEvent.OutputObject
  >;
  getEvent(
    key: "DIDUpdated"
  ): TypedContractEvent<
    DIDUpdatedEvent.InputTuple,
    DIDUpdatedEvent.OutputTuple,
    DIDUpdatedEvent.OutputObject
  >;
  getEvent(
    key: "KeyRotated"
  ): TypedContractEvent<
    KeyRotatedEvent.InputTuple,
    KeyRotatedEvent.OutputTuple,
    KeyRotatedEvent.OutputObject
  >;

  filters: {
    "ControllerRotated(string,address,address)": TypedContractEvent<
      ControllerRotatedEvent.InputTuple,
      ControllerRotatedEvent.OutputTuple,
      ControllerRotatedEvent.OutputObject
    >;
    ControllerRotated: TypedContractEvent<
      ControllerRotatedEvent.InputTuple,
      ControllerRotatedEvent.OutputTuple,
      ControllerRotatedEvent.OutputObject
    >;

    "DIDRegistered(string,address,string,uint256)": TypedContractEvent<
      DIDRegisteredEvent.InputTuple,
      DIDRegisteredEvent.OutputTuple,
      DIDRegisteredEvent.OutputObject
    >;
    DIDRegistered: TypedContractEvent<
      DIDRegisteredEvent.InputTuple,
      DIDRegisteredEvent.OutputTuple,
      DIDRegisteredEvent.OutputObject
    >;

    "DIDUpdated(string,string)": TypedContractEvent<
      DIDUpdatedEvent.InputTuple,
      DIDUpdatedEvent.OutputTuple,
      DIDUpdatedEvent.OutputObject
    >;
    DIDUpdated: TypedContractEvent<
      DIDUpdatedEvent.InputTuple,
      DIDUpdatedEvent.OutputTuple,
      DIDUpdatedEvent.OutputObject
    >;

    "KeyRotated(string,bytes,bytes)": TypedContractEvent<
      KeyRotatedEvent.InputTuple,
      KeyRotatedEvent.OutputTuple,
      KeyRotatedEvent.OutputObject
    >;
    KeyRotated: TypedContractEvent<
      KeyRotatedEvent.InputTuple,
      KeyRotatedEvent.OutputTuple,
      KeyRotatedEvent.OutputObject
    >;
  };
}